Fixed gcc warning.

+++ src/sbbs3/js_queue.c
@@ -272,7 +272,7 @@ js_queue_constructor(JSContext *cx, JSObject *obj, uintN argc, jsval *argv, jsva
 {
 	uintN			argn=0;
 	char*			name=NULL;
-	long			flags=MSG_QUEUE_BIDIR;
+	int32			flags=MSG_QUEUE_BIDIR;
 	msg_queue_t*	q=NULL;
 	list_node_t*	n;
